rearrangement + initial
