+++ simple-bench.js
@@ -13,13 +13,14 @@ const perf_hooks = require('perf_hooks');
 const {performance: perf, PerformanceObserver: PerfObserver} = perf_hooks;
 const util = require('util');
 
+const definitions = require('./benchmark/definitions');
 const {
   configure,
   setup,
   groupSetup,
   tests,
   final
-} = require('./benchmark/test-definitions');
+} = definitions;
 
 if (!tests.noop) {
   tests.noop = async s => s;
@@ -138,8 +139,8 @@ obs.observe({entryTypes: ['measure', 'gc'], buffered: true});
 //
 async function test() {
   // call the tester's setup
-  if (setup) {
-    await (async() => setup(config))();
+  if (definitions.setup) {
+    await (async() => definitions.setup(config))();
   }
   if (groupSetup) {
     await (async() => groupSetup(config))();
@@ -194,7 +195,7 @@ test().then(() => {
   obs.disconnect();
   const gTimes = groupTimes.slice();
   const gcStats = {gcCounts, totalGCTime};
-  const data = {gTimes, gcStats, config};
+  const data = {gTimes, gcStats, stddevRange, config};
   if (memCheck) {
     data.memory = memory;
   }
